requirements.txt
asgiref==3.2.10
Django==3.0.8
pytz==2020.1
sqlparse==0.3.1
djangorestframework
markdown
django-filter